LOOP-1380: improved loop notification settings
This propagates the "warning icon" all the way up to the "Notifications" button in Settings.
This took a bit of trickery to publish to old-style UI.
